Move EditAliasesForm
